requirements.txt
django
djangorestframework
djangorestframework-jwt
django-logpipe

